Moved hotel tests to hotel.entities.hotel Added hotelCheckinWithInvalidConfirmationNumber
